FPLA-720 moved things util class RoboticsDataVerificationHelper into utils package
